feat(backoffice-v2): update dependencies and add new components
- Add BusinessReportSectionsObserver for report navigation
- Introduce BusinessReportOptionsDropdown for report options
- Include hooks for PDF generation and report reference management
