+++ apps/backoffice-v2/package.json
@@ -54,10 +54,10 @@
   "dependencies": {
     "@ballerine/blocks": "0.2.35-demo-200225.0",
     "@ballerine/common": "0.9.76-demo-200225.0",
-    "@ballerine/workflow-browser-sdk": "0.6.96-demo-200225.0",
-    "@ballerine/workflow-node-sdk": "0.6.96-demo-200225.0",
     "@ballerine/react-pdf-toolkit": "^1.2.78-demo-200225.0",
     "@ballerine/ui": "^0.5.78-demo-200225.0",
+    "@ballerine/workflow-browser-sdk": "0.6.96-demo-200225.0",
+    "@ballerine/workflow-node-sdk": "0.6.96-demo-200225.0",
     "@botpress/webchat": "^2.1.10",
     "@botpress/webchat-generator": "^0.2.9",
     "@fontsource/inter": "^4.5.15",
@@ -113,6 +113,7 @@
     "eslint-plugin-tailwindcss": "^3.8.0",
     "face-api.js": "^0.22.2",
     "framer-motion": "^8.3.4",
+    "html2canvas": "^1.4.1",
     "i18next": "^22.4.9",
     "i18next-browser-languagedetector": "^7.0.1",
     "i18next-http-backend": "^2.1.1",
@@ -140,6 +141,7 @@
     "react-leaflet": "^4.2.1",
     "react-medium-image-zoom": "^5.2.10",
     "react-router-dom": "^6.11.2",
+    "react-to-pdf": "^1.0.1",
     "react-zoom-pan-pinch": "^3.0.8",
     "recharts": "^2.7.2",
     "sonner": "^1.4.3",
