+++ apps/backoffice-v2/src/pages/MerchantMonitoringBusinessReport/MerchantMonitoringBusinessReport.page.tsx
@@ -1,50 +1,31 @@
-import { MERCHANT_REPORT_STATUSES_MAP, UPDATEABLE_REPORT_STATUSES } from '@ballerine/common';
 import {
-  ContentTooltip,
   Dialog,
   DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
   DialogTrigger,
-  DropdownMenu,
-  DropdownMenuContent,
   DropdownMenuItem,
-  DropdownMenuTrigger,
   Skeleton,
-  TextArea,
   TextWithNAFallback,
 } from '@ballerine/ui';
 import dayjs from 'dayjs';
-import { ArrowLeft, ArrowRightIcon, ChevronLeft, FileQuestion } from 'lucide-react';
-import React, { forwardRef, FunctionComponent } from 'react';
 import { Link } from 'react-router-dom';
+import React, { forwardRef, FunctionComponent } from 'react';
+import { ArrowLeft, ArrowRightIcon, ChevronLeft, FileQuestion } from 'lucide-react';
+import { MERCHANT_REPORT_STATUSES_MAP, UPDATEABLE_REPORT_STATUSES } from '@ballerine/common';
 
-import { Button } from '@/common/components/atoms/Button/Button';
+import { ctw } from '@/common/utils/ctw/ctw';
+import { Notes } from '@/domains/notes/Notes';
 import { Card } from '@/common/components/atoms/Card/Card';
-import { CardContent } from '@/common/components/atoms/Card/Card.Content';
+import { BALLERINE_CALENDLY_LINK } from '@/common/constants';
+import { Button } from '@/common/components/atoms/Button/Button';
+import { CardTitle } from '@/common/components/atoms/Card/Card.Title';
 import { CardFooter } from '@/common/components/atoms/Card/Card.Footer';
 import { CardHeader } from '@/common/components/atoms/Card/Card.Header';
-import { CardTitle } from '@/common/components/atoms/Card/Card.Title';
-import { Select } from '@/common/components/atoms/Select/Select';
-import { SelectContent } from '@/common/components/atoms/Select/Select.Content';
-import { SelectItem } from '@/common/components/atoms/Select/Select.Item';
-import { SelectTrigger } from '@/common/components/atoms/Select/Select.Trigger';
-import { SelectValue } from '@/common/components/atoms/Select/Select.Value';
+import { CardContent } from '@/common/components/atoms/Card/Card.Content';
 import { Separator } from '@/common/components/atoms/Separator/Separator';
+import { BusinessReportOptionsDropdown } from './BusinessReportOptionsDropdown';
 import { NotesButton } from '@/common/components/molecules/NotesButton/NotesButton';
-import { Form } from '@/common/components/organisms/Form/Form';
-import { FormControl } from '@/common/components/organisms/Form/Form.Control';
-import { FormField } from '@/common/components/organisms/Form/Form.Field';
-import { FormItem } from '@/common/components/organisms/Form/Form.Item';
-import { FormLabel } from '@/common/components/organisms/Form/Form.Label';
-import { FormMessage } from '@/common/components/organisms/Form/Form.Message';
 import { SidebarInset, SidebarProvider } from '@/common/components/organisms/Sidebar/Sidebar';
-import { BALLERINE_CALENDLY_LINK } from '@/common/constants';
-import { ctw } from '@/common/utils/ctw/ctw';
 import { BusinessReport } from '@/domains/business-reports/components/BusinessReport/BusinessReport';
-import { Notes } from '@/domains/notes/Notes';
 import { MerchantMonitoringReportStatus } from '@/pages/MerchantMonitoring/components/MerchantMonitoringReportStatus/MerchantMonitoringReportStatus';
 import { useMerchantMonitoringBusinessReportLogic } from '@/pages/MerchantMonitoringBusinessReport/hooks/useMerchantMonitoringBusinessReportLogic/useMerchantMonitoringBusinessReportLogic';
 
@@ -80,175 +61,6 @@ export const DialogDropdownItem = forwardRef<
 
 DialogDropdownItem.displayName = 'DialogDropdownItem';
 
-const BusinessReportOptionsDropdown: FunctionComponent<
-  Pick<
-    ReturnType<typeof useMerchantMonitoringBusinessReportLogic>,
-    | 'isDropdownOpen'
-    | 'setIsDropdownOpen'
-    | 'isDeboardModalOpen'
-    | 'setIsDeboardModalOpen'
-    | 'isDemoAccount'
-    | 'businessReport'
-    | 'turnOngoingMonitoringOn'
-    | 'form'
-    | 'onSubmit'
-    | 'deboardingReasonOptions'
-  >
-> = ({
-  isDropdownOpen,
-  setIsDropdownOpen,
-  isDeboardModalOpen,
-  setIsDeboardModalOpen,
-  isDemoAccount,
-  businessReport,
-  turnOngoingMonitoringOn,
-  form,
-  onSubmit,
-  deboardingReasonOptions,
-}) => {
-  return (
-    <DropdownMenu open={isDropdownOpen} onOpenChange={setIsDropdownOpen} modal={false}>
-      <DropdownMenuTrigger asChild>
-        <Button
-          variant="outline"
-          className={
-            'ml-auto px-2 py-0 text-xs aria-disabled:pointer-events-none aria-disabled:opacity-50'
-          }
-        >
-          Options
-        </Button>
-      </DropdownMenuTrigger>
-
-      <DropdownMenuContent
-        align="end"
-        onEscapeKeyDown={e => {
-          if (isDeboardModalOpen) {
-            e.preventDefault();
-            setIsDeboardModalOpen(false);
-          }
-        }}
-      >
-        <ContentTooltip
-          description={
-            <p>
-              This feature is not available for trial accounts.
-              <br />
-              Talk to us to get full access
-            </p>
-          }
-          props={{
-            tooltipContent: {
-              className: ctw({ hidden: !isDemoAccount }),
-            },
-          }}
-        >
-          {businessReport?.monitoringStatus === true ? (
-            <DialogDropdownItem
-              triggerChildren={
-                <Button variant={'ghost'} className="justify-start">
-                  Turn Monitoring Off
-                </Button>
-              }
-              open={isDeboardModalOpen}
-              onOpenChange={setIsDeboardModalOpen}
-              disabled={isDemoAccount}
-            >
-              <DialogHeader>
-                <DialogTitle>Confirm Deboarding</DialogTitle>
-                <DialogDescription>
-                  Are you sure you want to deboard this merchant (turn the monitoring off)?
-                </DialogDescription>
-              </DialogHeader>
-
-              <Form {...form}>
-                <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-                  <FormField
-                    control={form.control}
-                    name="reason"
-                    render={({ field }) => (
-                      <FormItem>
-                        <Select onValueChange={field.onChange} value={field.value}>
-                          <FormLabel>Reason</FormLabel>
-
-                          <FormControl>
-                            <SelectTrigger className="h-9 w-full border-input p-1 shadow-sm">
-                              <SelectValue placeholder="Select a reason" />
-                            </SelectTrigger>
-                          </FormControl>
-                          <FormMessage />
-                          <SelectContent>
-                            {deboardingReasonOptions?.map((option, index) => {
-                              return (
-                                <SelectItem key={index} value={option}>
-                                  {option}
-                                </SelectItem>
-                              );
-                            })}
-                          </SelectContent>
-                        </Select>
-                      </FormItem>
-                    )}
-                  />
-
-                  <FormField
-                    control={form.control}
-                    name="userReason"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Additional details</FormLabel>
-
-                        <FormControl>
-                          <TextArea {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-
-                  <DialogFooter className="mt-6 flex justify-end space-x-4">
-                    <Button
-                      type="button"
-                      onClick={() => {
-                        setIsDeboardModalOpen(false);
-                      }}
-                      variant="ghost"
-                    >
-                      Cancel
-                    </Button>
-                    <Button type="submit" variant="destructive">
-                      Turn Off
-                    </Button>
-                  </DialogFooter>
-                </form>
-              </Form>
-            </DialogDropdownItem>
-          ) : (
-            <Button
-              onClick={() => {
-                if (!businessReport?.business.id) {
-                  throw new Error('Business ID is missing');
-                }
-
-                turnOngoingMonitoringOn(businessReport.business.id, {
-                  onSuccess: () => {
-                    setIsDeboardModalOpen(false);
-                    setIsDropdownOpen(false);
-                  },
-                });
-              }}
-              variant={'ghost'}
-              className="justify-start disabled:bg-inherit disabled:text-foreground"
-              disabled={isDemoAccount}
-            >
-              Turn Monitoring On
-            </Button>
-          )}
-        </ContentTooltip>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-};
-
 export const MerchantMonitoringBusinessReport: FunctionComponent = () => {
   const {
     onNavigateBack,
@@ -259,15 +71,7 @@ export const MerchantMonitoringBusinessReport: FunctionComponent = () => {
     isFetchingBusinessReport,
     locale,
     isDemoAccount,
-
-    // turnOngoingMonitoringOn,
-    // isDeboardModalOpen,
-    // setIsDeboardModalOpen,
-    // isDropdownOpen,
-    // setIsDropdownOpen,
-    // form,
-    // onSubmit,
-    // deboardingReasonOptions,
+    reportRef,
     ...dropdownProps
   } = useMerchantMonitoringBusinessReportLogic();
 
@@ -359,68 +163,70 @@ export const MerchantMonitoringBusinessReport: FunctionComponent = () => {
           we can easily determine negative margin and width required to properly display the separator. */}
           <Separator className="-ml-6 mb-4 w-[calc(100%+3rem)]" />
 
-          {isFetchingBusinessReport ? (
-            <Skeleton className="h-6 w-32" />
-          ) : (
-            <div className="flex items-center justify-between">
-              <TextWithNAFallback as={'h2'} className="pb-4 text-2xl font-bold">
-                {websiteWithNoProtocol}
-              </TextWithNAFallback>
+          <div ref={reportRef}>
+            {isFetchingBusinessReport ? (
+              <Skeleton className="h-6 w-32" />
+            ) : (
+              <div className="flex items-center justify-between">
+                <TextWithNAFallback as={'h2'} className="pb-4 text-2xl font-bold">
+                  {websiteWithNoProtocol}
+                </TextWithNAFallback>
 
-              {isDemoAccount && (
-                <BusinessReportOptionsDropdown
-                  {...dropdownProps}
-                  businessReport={businessReport}
-                  isDemoAccount={isDemoAccount}
-                />
-              )}
-            </div>
-          )}
-          {isFetchingBusinessReport ? (
-            <Skeleton className="my-6 h-6 w-2/3" />
-          ) : (
-            <div className={`flex items-center space-x-8 pb-4`}>
-              <div className={`flex items-center`}>
-                <span className={`me-4 text-sm leading-6 text-slate-400`}>Status</span>
-                <MerchantMonitoringReportStatus
-                  reportId={businessReport?.id}
-                  status={businessReport?.status}
-                  businessId={businessReport?.business.id}
-                />
+                {isDemoAccount && (
+                  <BusinessReportOptionsDropdown
+                    {...dropdownProps}
+                    businessReport={businessReport}
+                    isDemoAccount={isDemoAccount}
+                  />
+                )}
               </div>
-              <div className={`text-sm`}>
-                <span className={`me-2 leading-6 text-slate-400`}>Created at</span>
-                {businessReport?.displayDate &&
-                  dayjs(new Date(businessReport?.displayDate)).format('HH:mm MMM Do, YYYY')}
+            )}
+            {isFetchingBusinessReport ? (
+              <Skeleton className="my-6 h-6 w-2/3" />
+            ) : (
+              <div className={`flex items-center space-x-8 pb-4`}>
+                <div className={`flex items-center`}>
+                  <span className={`me-4 text-sm leading-6 text-slate-400`}>Status</span>
+                  <MerchantMonitoringReportStatus
+                    reportId={businessReport?.id}
+                    status={businessReport?.status}
+                    businessId={businessReport?.business.id}
+                  />
+                </div>
+                <div className={`text-sm`}>
+                  <span className={`me-2 leading-6 text-slate-400`}>Created at</span>
+                  {businessReport?.displayDate &&
+                    dayjs(new Date(businessReport?.displayDate)).format('HH:mm MMM Do, YYYY')}
+                </div>
+                <div className={`flex items-center space-x-2 text-sm`}>
+                  <span className={`text-slate-400`}>Monitoring Status</span>
+                  <span
+                    className={ctw('select-none rounded-full d-3', {
+                      'bg-success': businessReport?.monitoringStatus,
+                      'bg-slate-400': !businessReport?.monitoringStatus,
+                    })}
+                  >
+                    &nbsp;
+                  </span>
+                </div>
+                <NotesButton numberOfNotes={notes?.length} />
               </div>
-              <div className={`flex items-center space-x-2 text-sm`}>
-                <span className={`text-slate-400`}>Monitoring Status</span>
-                <span
-                  className={ctw('select-none rounded-full d-3', {
-                    'bg-success': businessReport?.monitoringStatus,
-                    'bg-slate-400': !businessReport?.monitoringStatus,
-                  })}
-                >
-                  &nbsp;
-                </span>
-              </div>
-              <NotesButton numberOfNotes={notes?.length} />
-            </div>
-          )}
-          {isFetchingBusinessReport || !businessReport ? (
-            <>
-              <Skeleton className="h-6 w-72" />
-              <Skeleton className="mt-6 h-4 w-40" />
+            )}
+            {isFetchingBusinessReport || !businessReport ? (
+              <>
+                <Skeleton className="h-6 w-72" />
+                <Skeleton className="mt-6 h-4 w-40" />
 
-              <div className="mt-6 flex h-[24rem] w-full flex-nowrap gap-8">
-                <Skeleton className="w-2/3" />
-                <Skeleton className="w-1/3" />
-              </div>
-              <Skeleton className="mt-6 h-[16rem]" />
-            </>
-          ) : (
-            <BusinessReport report={businessReport} />
-          )}
+                <div className="mt-6 flex h-[24rem] w-full flex-nowrap gap-8">
+                  <Skeleton className="w-2/3" />
+                  <Skeleton className="w-1/3" />
+                </div>
+                <Skeleton className="mt-6 h-[16rem]" />
+              </>
+            ) : (
+              <BusinessReport report={businessReport} />
+            )}
+          </div>
         </section>
       </SidebarInset>
       <Notes
